Figured out solution to keeping relative coordinates the same
When resizing, the pane does not layout it's children. Thus after resizing, the children may not be over the same place as originally. This will not work since the circles and lines must stay over the same place in the image no matter what. Thus I finally used low level binding api to bind the circle x and y to a calculated and correct new x and y based on the initial bounds of the image. The initial x and y will be supplied by added vertices. A better solution now apparent to me, is to extend the Pane class and manage the children's layout since this is a problem of layout, not a problem of where to position.